+++ web/src/features/fireWatch/components/FireWatchDashboard.tsx
@@ -1,43 +1,14 @@
 import { AppDispatch } from '@/app/store'
 import DetailPanelContent from '@/features/fireWatch/components/DetailPanelContent'
 import FireWatchDetailsModal from '@/features/fireWatch/components/FireWatchDetailsModal'
-import {
-  BurnForecast,
-  BurnStatusEnum,
-  FireWatch,
-  FireWatchBurnForecast,
-  FuelTypeEnum,
-  PrescriptionEnum
-} from '@/features/fireWatch/interfaces'
+import { BurnWatchRow, FireWatchBurnForecast } from '@/features/fireWatch/interfaces'
 import { fetchBurnForecasts, selectBurnForecasts } from '@/features/fireWatch/slices/burnForecastSlice'
 import InfoIcon from '@mui/icons-material/Info'
 import { Box, styled, Typography, useTheme } from '@mui/material'
-import {
-  DataGridPro,
-  DataGridProProps,
-  GridActionsCellItem,
-  GridColDef,
-  GridValueFormatterParams
-} from '@mui/x-data-grid-pro'
-import { isNull } from 'lodash'
-import { DateTime } from 'luxon'
+import { DataGridPro, DataGridProProps, GridActionsCellItem, GridColDef } from '@mui/x-data-grid-pro'
 import React, { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { FireWatchPrescriptionColors } from 'app/theme'
-
-export interface BurnWatchRow {
-  id: number
-  title: string
-  fireCentre: string
-  station: string
-  fuelType: FuelTypeEnum
-  status: BurnStatusEnum
-  burnWindowStart: DateTime | null
-  burnWindowEnd: DateTime | null
-  inPrescription: PrescriptionEnum
-  fireWatch: FireWatch
-  burnForecasts: BurnForecast[]
-}
 
 const StyledDataGrid = styled(DataGridPro)(({ theme }) => ({
   ['&.in-prescription-all']: { backGroundColor: theme.palette.success },
