+++ web/src/features/fireWatch/constants.ts
@@ -1,0 +1,1 @@
+export const FORM_MAX_WIDTH = 768
